refactor(wms): place cli in subfolder for easy install
`go install github.com/jordangarrison/whats-my-status/wms@latest`.

BREAKING CHANGE: You have to install from a new path.
