+++ Taskfile.yml
@@ -22,24 +22,29 @@ tasks:
 
   # Architecture specific build tasks
   build_linux_amd64:
+    dir: ./wms
     cmds:
-      - env GOOS=linux GOARCH=amd64 go build -v -o build/{{.PACKAGE_NAME}}-linux-amd64-{{.VERSION}}
+      - env GOOS=linux GOARCH=amd64 go build -v -o ../build/{{.PACKAGE_NAME}}-linux-amd64-{{.VERSION}}
 
   build_linux_arm:
+    dir: ./wms
     cmds:
-      - env GOOS=linux GOARCH=arm go build -v -o build/{{.PACKAGE_NAME}}-linux-arm-{{.VERSION}}
+      - env GOOS=linux GOARCH=arm go build -v -o ../build/{{.PACKAGE_NAME}}-linux-arm-{{.VERSION}}
 
   build_linux_arm64:
+    dir: ./wms
     cmds:
-      - env GOOS=linux GOARCH=arm64 go build -v -o build/{{.PACKAGE_NAME}}-linux-arm64-{{.VERSION}}
+      - env GOOS=linux GOARCH=arm64 go build -v -o ../build/{{.PACKAGE_NAME}}-linux-arm64-{{.VERSION}}
 
   build_darwin_amd64:
+    dir: ./wms
     cmds:
-      - env GOOS=darwin GOARCH=amd64 go build -v -o build/{{.PACKAGE_NAME}}-darwin-amd64-{{.VERSION}}
+      - env GOOS=darwin GOARCH=amd64 go build -v -o ../build/{{.PACKAGE_NAME}}-darwin-amd64-{{.VERSION}}
 
   build_windows_amd64:
+    dir: ./wms
     cmds:
-      - env GOOS=windows GOARCH=amd64 go build -v -o build/{{.PACKAGE_NAME}}-windows-amd64-{{.VERSION}}.exe
+      - env GOOS=windows GOARCH=amd64 go build -v -o ../build/{{.PACKAGE_NAME}}-windows-amd64-{{.VERSION}}.exe
 
   # Clean the build directory
   clean:
@@ -75,7 +80,7 @@ tasks:
       - go mod tidy
   test:
     cmds:
-      - go vet
+      - go vet ./...
       - go test -count=1 -v ./...
 
   # Publish the package
