+++ wms/cmd/status.go
